+++ src/Lambda/Notebook/Kernel/Model.hs
@@ -1,5 +1,7 @@
 {-# LANGUAGE DeriveAnyClass #-}
 {-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE GADTs #-}
 
 module Lambda.Notebook.Kernel.Model
   ( Kernel (..),